Try removing query string when looking for redirects.
If a redirect URL has a query at the end, eg. /foo?bar=quux, it can now be
routed properly as /foo. We lookup the redirect with the query first, so
one can create redirects with specific query requirements. However, if
we cant find one, we search again without the query string, which is usually
better than just giving up entirely.

In most cases, the query is just analytics parameters like orig_ql or utm_*.